socketio_live/requirements.txt
bidict==0.22.0
click==8.1.3
dnspython==2.2.1
eventlet==0.33.1
Flask==2.2.2
flask-sock==0.5.2
Flask-SocketIO==5.3.1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-engineio==4.3.4
python-socketio==5.7.1
simple-websocket==0.8.1
six==1.16.0
typing_extensions==4.3.0
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.8.1
